Refactored models project structure with TYPE_CHECKING
