+++ labelapp/labeling_app/app/models/__init__.py
@@ -1,0 +1,6 @@
+from app.models.extras import LabelModel, UserProjectModel
+from app.models.projects import ProjectModel
+from app.models.urls import UrlModel
+from app.models.users import UserModel
+
+__all__ = [ProjectModel, UserModel, UrlModel, LabelModel, UserProjectModel]
