--- labelapp/labeling_app/app/schemas/read.py
@@ -1,9 +1,0 @@
-import uuid
-
-from app.models import LabelModel
-from app.schemas.projects import ProjectBase
-
-
-class ProjectRead(ProjectBase):
-    owner_id: uuid.UUID
-    labels: list[LabelModel] = []
